Add: 버튼 클릭 시 페이지 이동 기능 추가
